Move Maven package installer to reusable core packages
